requirements.txt
click>=8.0
anthropic>=0.40
fastmcp>=0.1
unidiff>=0.7
pyyaml>=6.0
python-dotenv>=1.0
